+++ wp-content/plugins/object-cache-pro/src/Plugin/Licensing.php
@@ -194,7 +194,7 @@ trait Licensing
     /**
      * Fetch the license for configured token.
      *
-     * @return object|\WP_Error
+     * @return \RedisCachePro\Support\PluginApiLicenseResponse|\WP_Error
      */
     protected function fetchLicense()
     {
@@ -204,9 +204,13 @@ trait Licensing
             return new WP_Error('objectcache_fetch_failed', sprintf(
                 'Could not verify license. %s',
                 $response->get_error_message()
-            ), ['token' => $this->token()]);
+            ), array_merge(
+                ['token' => $this->token()],
+                $response->get_error_data() ?? []
+            ));
         }
 
+        /** @var \RedisCachePro\Support\PluginApiLicenseResponse $response */
         return $response;
     }
 
@@ -298,17 +302,26 @@ trait Licensing
         ]);
 
         if (is_wp_error($response)) {
-            return $response;
+            return new WP_Error(
+                'objectcache_http_error',
+                $response->get_error_message(),
+                ['code' => $response->get_error_code()]
+            );
         }
 
         $body = wp_remote_retrieve_body($response);
+        $headers = wp_remote_retrieve_headers($response);
         $status = wp_remote_retrieve_response_code($response);
 
         if ($status >= 400) {
             return new WP_Error(
                 'objectcache_api_error',
                 "Request returned status code {$status}",
-                ['status' => $status]
+                [
+                    'status' => $status,
+                    'cf-ray' => $headers['cf-ray'] ?? null,
+                    'cf-mitigated' => $headers['cf-mitigated'] ?? null,
+                ]
             );
         }
 
@@ -316,7 +329,11 @@ trait Licensing
             return new WP_Error(
                 'objectcache_api_error',
                 'Request returned empty body',
-                ['status' => $status]
+                [
+                    'status' => $status,
+                    'cf-ray' => $headers['cf-ray'] ?? null,
+                    'cf-mitigated' => $headers['cf-mitigated'] ?? null,
+                ]
             );
         }
 
@@ -338,17 +355,24 @@ trait Licensing
     /**
      * Performs a `plugin/info` request and returns the result.
      *
-     * @return \RedisCachePro\Support\PluginApiResponse|\WP_Error
+     * @return \RedisCachePro\Support\PluginApiInfoResponse|\WP_Error
      */
     public function pluginInfoRequest()
     {
-        return $this->request('plugin/info');
+        $response = $this->request('plugin/info');
+
+        if (is_wp_error($response)) {
+            return $response;
+        }
+
+        /** @var \RedisCachePro\Support\PluginApiInfoResponse $response */
+        return $response;
     }
 
     /**
      * Performs a `plugin/update` request and returns the result.
      *
-     * @return \RedisCachePro\Support\PluginApiResponse|\WP_Error
+     * @return \RedisCachePro\Support\PluginApiUpdateResponse|\WP_Error
      */
     public function pluginUpdateRequest()
     {
@@ -358,6 +382,7 @@ trait Licensing
             return $response;
         }
 
+        /** @var \RedisCachePro\Support\PluginApiUpdateResponse $response */
         set_site_transient('objectcache_update', (object) [
             'version' => $response->version,
             'last_check' => time(),