+++ wp-content/themes/livability/template-parts/content/content-places-v2.php
@@ -42,6 +42,12 @@ $topic_args = array(
    }
 
    $is_2024_bp = has_term('2024', 'best_places_years');
+   $parent_obj = get_post_parent();
+   $state_name = $parent_obj->post_title;
+   $state_abbv = $parent_obj->post_name;
+   $city_name = substr(get_the_title($place[0]), 0, -4); 
+//    print_r($parent_obj);
+
 ?>
 
 <article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
@@ -52,14 +58,14 @@ $topic_args = array(
                 <?php get_template_part( 'template-parts/blocks/ad-one' ); ?>
             </div>
         </div>
-
 <div class="place-column__parent">
     <div class="place-column__nav">
     <ul class="place-side-nav">
-        
+        <?php if ($is_2024_bp): ?>
         <li><a href="#top-100" >2024 Top 100</a></li>
         <li><a href="#things-to-do">Things to Do</a></li>
         <li><a href="#economy">Economy</a></li>
+        <?php endif; ?>
         <li><a href="#brands-to-know">Brands to Know</a></li>
         <li><a href="#quick-facts">Quick Facts</a></li>
         <li><a href="#weather">Weather</a></li>
@@ -70,40 +76,18 @@ $topic_args = array(
         echo array_key_exists('healthy-places', $topics_array) ? '<li><a href="#healthy-places">Healthy Places</a></li>': '';
         echo array_key_exists('make-your-move', $topics_array) ? '<li><a href="#make-your-move">Make Your Move</a></li>': '';
         echo array_key_exists('where-to-live-now', $topics_array) ? '<li><a href="#where-to-live-now">Where to Live Now</a></li>': '';
-        echo array_key_exists('education-careers-opportunity', $topics_array) ? '<li><a href="#education-careers-and-opportunities">Education, Careers & Opportunities</a></li>': '';
+        echo array_key_exists('education-careers-opportunity', $topics_array) ? '<li><a href="#education-careers-and-opportunity">Education, Careers & Opportunities</a></li>': '';
         echo array_key_exists('love-where-you-live', $topics_array) ? '<li><a href="#love-where-you-live">Love Where You Live</a></li>': '';
         echo array_key_exists('where-to-live-now', $topics_array) ? '<li><a href="#more-about-living-in">More About Living in </a></li>': '';
         ?>
+        <li><a href="#more-about-living">More About Living in <?php echo $state_name;   ?></a></li>
 
     </ul>
     </div>
     <div class="place-column__content">
     
-        <div class="wp-block-columns liv-columns">
+        <div class="wp-block-columns liv-column-p">
             <div class="wp-block-column">
-                <div class="wp-block-columns liv-columns-2">
-                    <div class="wp-block-column"></div>
-                    <div class="wp-block-column">
-                        <div id="crumbs">
-                            <?php echo return_breadcrumbs(); ?> 
-                        </div>
-                        <?php if (has_term('2024', 'best_places_years')) {echo 'has term';} else {echo 'not term';} 
-                        // print_r(get_terms()); 
-                        echo get_the_term_list(get_the_ID(),'best_places_years' );?>
-                     
-                    </div>
-                </div>
-                <div class="wp-block-columns liv-columns-2">
-                    <div class="wp-block-column">
-
-                    <?php 
-                    //get_template_part( 'template-parts/blocks/top-one-hundred-link'); ?>
-                    
-                    </div>
-                    <div class="wp-block-column">
-
-                    
-                     
                      <?php 
                      if ($is_2024_bp) {
                          echo '<h2 class="h1">'.get_the_title().'</h2>';
@@ -122,35 +106,71 @@ $topic_args = array(
                         get_template_part( 'template-parts/blocks/madlib' ); 
                     }
                     get_template_part( 'template-parts/blocks/embedded-cc-article' );
-                    get_template_part( 'template-parts/blocks/link-to-2023-best-place' );
+                    // get_template_part( 'template-parts/blocks/link-to-2023-best-place' );
                      ?>
                     </div>
-                </div>
-            </div>
+                <!-- </div> -->
+            <!-- </div> -->
             <div class="wp-block-column">
-            <?php get_template_part( 'template-parts/blocks/ad-two' ); ?>
+            <?php 
+            get_template_part( 'template-parts/blocks/ad-two' ); ?>
             </div>
         </div>
         <div class="wp-block-columns">
             <div class="wp-block-column">
-            <?php get_template_part( 'template-parts/blocks/quick-facts' ); ?>
-        </div></div>
+            <?php get_template_part( 'template-parts/blocks/brand-stories', null, array('city' => $city_name, 'state' => $state_name) );
+            get_template_part( 'template-parts/blocks/quick-facts-2', null, array('city' => $city_name, 'state' => $state_name ) ); ?>
+            </div>
+        </div>
         
-        <?php get_template_part( 'template-parts/blocks/city-map' ); ?>
+        <?php echo '<h2 id="map">Map of '.$city_name.', '.$state_name.'</h2>'; 
+        get_template_part( 'template-parts/blocks/city-map' ); ?>
         
         <div class="wp-block-colomns">
             <div class="wp-block-column">
-      
-            <?php get_template_part( 'template-parts/blocks/place-brand-story' ); ?>
-            <?php get_template_part( 'template-parts/blocks/place-topics',null, array('category' => 'experiences-adventures') ); ?>
-            <?php get_template_part( 'template-parts/blocks/ad-three' ); ?>
-            <?php get_template_part( 'template-parts/blocks/place-topics',null, array('category' => 'food-scenes') ); ?>
-            <?php get_template_part( 'template-parts/blocks/place-topics',null, array('category' => 'healthy-places') ); ?>
-            <?php get_template_part( 'template-parts/blocks/place-topics',null, array('category' => 'affordable-places-to-live') ); ?>
-            <?php get_template_part( 'template-parts/blocks/place-topics',null, array('category' => 'make-your-move') ); ?>
-            <?php get_template_part( 'template-parts/blocks/place-topics',null, array('category' => 'where-to-live-now') ); ?>
-            <?php get_template_part( 'template-parts/blocks/place-topics',null, array('category' => 'education-careers-opportunity') ); ?>
-            <?php get_template_part( 'template-parts/blocks/place-topics',null, array('category' => 'love-where-you-live') ); ?>
+            <?php 
+            if (array_key_exists('experiences-adventures', $topics_array)) {
+                echo '<h2 id="experiences-adventures">Experiences & Adventures in '.$city_name.', '.$state_name.'</h2>';
+                
+                get_template_part( 'template-parts/blocks/place-topics-2', null, array('posts' => $topics_array['experiences-adventures']));
+            }
+
+            if (array_key_exists('food-scenes', $topics_array)) {
+                echo '<h2 id="food-scenes">Food Scenes in '.$city_name.', '.$state_name.'</h2>';
+                
+                get_template_part( 'template-parts/blocks/place-topics-2', null, array('posts' => $topics_array['food-scenes']));
+            }  
+            
+            if (array_key_exists('healthy-places', $topics_array)) {
+                echo '<h2 id="healthy-places">Healthy Places in '.$city_name.', '.$state_name.'</h2>';
+                
+                get_template_part( 'template-parts/blocks/place-topics-2', null, array('posts' => $topics_array['healthy-places']));
+            }   
+            
+            if (array_key_exists('make-your-move', $topics_array)) {
+                echo '<h2 id="make-your-move">Make Your Move to '.$city_name.', '.$state_name.'</h2>';
+                
+                get_template_part( 'template-parts/blocks/place-topics-2', null, array('posts' => $topics_array['make-your-move']));
+            }    
+            
+            if (array_key_exists('where-to-live-now', $topics_array)) {
+                echo '<h2 id="where-to-live-now">'.$city_name.', '.$state_name.': Where to Live Now</h2>';
+                
+                get_template_part( 'template-parts/blocks/place-topics-2', null, array('posts' => $topics_array['where-to-live-now']));
+            }               
+
+            if (array_key_exists('education-careers-opportunity', $topics_array)) {
+                echo '<h2 id="education-careers-and-opportunity">Education, Careers and Opportunity in '.$city_name.', '.$state_name.'</h2>';
+                
+                get_template_part( 'template-parts/blocks/place-topics-2', null, array('posts' => $topics_array['education-careers-opportunity']));
+            }
+
+            if (array_key_exists('love-where-you-live', $topics_array)) {
+                echo '<h2 id="love-where-you-live">'.$city_name.', '.$state_name.': Love Where You Live</h2>';
+                
+                get_template_part( 'template-parts/blocks/place-topics-2', null, array('posts' => $topics_array['love-where-you-live']));
+            }            
+            ?>
             </div>
         </div>
 
@@ -159,7 +179,8 @@ $topic_args = array(
 </div><!-- place-column__parent --> 
 		
 	</div><!-- .entry-content -->
-
+    <?php get_template_part( 'template-parts/blocks/more-about-living', null, array('state' => $state_name, 'abbv' => $state_abbv) ); ?>
+